rsc_types: get rid of fw_custom and fixup fw_rsc_custom
get rid of fw_custom to make fw_rsc_custom more scalable.
The only format that is fixed in case of custom resource is
the header. i.e. type, sub_type and size.

The data will change depending on type of custom resource
and there is no point in defining a union for it.
Let the users populate the right kind of data just after
the fw_rsc_custom header.

As a reference we fix only one example.

TODO: fix all examples.

Signed-off-by: Roger Quadros <rogerq@ti.com>